Change the source of constants for 'Tags' to come from 'confetti-ds'. Refactor 'skin' controls of 'Tags' to be radio groups. Refactor interaction with 'thumbSrc' prop in 'Tag' stories. Refactor behaviour of restore button in 'RemovableTag'
