+++ stories/Tag/DropdownTag.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { TAG_ICONS, TAG_SKINS, TAG_COLORS } from "./constants";
+import { ICON_TYPES, TAG_COLORS } from "../../confetti-ds/src/constants";
 import { DropdownTag as Component  } from "../../confetti-ds/src";
 
 export default {
@@ -8,18 +8,19 @@ export default {
   component: Component,
   argTypes: {
     icon: {
-      control: { type: "select", options: TAG_ICONS },
+      control: { type: "select", options: ICON_TYPES },
     },
     color: {
       control: { type: "select", options: TAG_COLORS },
     },
     skin: {
-      control: { type: "select", options: TAG_SKINS },
+      control: { type: "inline-radio", options: ["pale", "vivid"] },
     },
   },
 };
 
 export const DropdownTag = (args) => <Component {...args} />;
+
 DropdownTag.args = {
   text: "demo tag",
 };
