+++ stories/Tag/RemovableTag.stories.js
@@ -1,23 +1,20 @@
 import React from "react";
 
-import { TAG_ICONS, TAG_THUMBS, TAG_SKINS, TAG_COLORS } from "./constants";
+import { ICON_TYPES, TAG_COLORS } from "../../confetti-ds/src/constants";
 import { Button, RemovableTag as Component } from "../../confetti-ds/src";
 
 export default {
   title: "Tag/Removable Tag",
   component: Component,
   argTypes: {
-    thumbSrc: {
-      control: { type: "select", options: TAG_THUMBS },
-    },
     icon: {
-      control: { type: "select", options: TAG_ICONS },
+      control: { type: "select", options: ICON_TYPES },
     },
     color: {
       control: { type: "select", options: TAG_COLORS },
     },
     skin: {
-      control: { type: "select", options: TAG_SKINS },
+      control: { type: "inline-radio", options: ["pale", "vivid"] },
     },
   },
 };
@@ -29,12 +26,12 @@ export const RemovableTag = (args) => {
     <React.Fragment>
       { showTag ?
         <Component {...args} onClick={() => setShowTag(false)} />
-        : null
-      }
+        :
+        <div>
+          <Button text="Restore tag" onClick={() => setShowTag(true)} size="small"/>
+        </div> }
 
-      <div style={{ marginTop: "16px" }}>
-        <Button text="Restore tag" onClick={() => setShowTag(true)} size="small"/>
-      </div>
+      <p>Use this link for a <strong>thumbSrc</strong> example: <a>http://avatars3.githubusercontent.com/u/1887591</a></p>
     </React.Fragment>
   )
 }
