+++ stories/Tag/SimpleTag.stories.js
@@ -1,28 +1,33 @@
 import React from "react";
 
-import { TAG_ICONS, TAG_THUMBS, TAG_SKINS, TAG_COLORS } from "./constants";
+import { ICON_TYPES, TAG_COLORS } from "../../confetti-ds/src/constants";
 import { SimpleTag as Component } from "../../confetti-ds/src";
 
 export default {
   title: "Tag/Simple Tag",
   component: Component,
   argTypes: {
-    thumbSrc: {
-      control: { type: "select", options: TAG_THUMBS },
-    },
     icon: {
-      control: { type: "select", options: TAG_ICONS },
+      control: { type: "select", options: ICON_TYPES },
     },
     color: {
       control: { type: "select", options: TAG_COLORS },
     },
     skin: {
-      control: { type: "select", options: TAG_SKINS },
+      control: { type: "inline-radio", options: ["pale", "vivid"] },
     },
   },
 };
 
-export const SimpleTag = (args) => <Component {...args} />;
+export const SimpleTag = (args) => {
+  return (
+    <React.Fragment>
+      <Component {...args} />
+      <p>Use this link for a <strong>thumbSrc</strong> example: <a>http://avatars3.githubusercontent.com/u/1887591</a></p>
+    </React.Fragment>
+  )
+};
+
 SimpleTag.args = {
   text: "demo tag",
 };
