+++ stories/Tag/TogglableTag.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { TAG_COLORS } from "./constants";
+import { TAG_COLORS } from "../../confetti-ds/src/constants";
 import { TogglableTag as Component } from "../../confetti-ds/src";
 
 export default {
