--- stories/Tag/constants.js
@@ -1,58 +1,0 @@
-export const TAG_ICONS = [
-  "",
-  "arrow-down",
-  "arrow-left",
-  "arrow-right",
-  "arrow-top",
-  "calendar",
-  "coin",
-  "collapse-closed",
-  "collapse-open",
-  "check",
-  "dropdown-closed",
-  "dropdown-open",
-  "edit",
-  "eye-closed",
-  "eye-opened",
-  "track",
-  "key",
-  "logout",
-  "magnifying-glass",
-  "minus",
-  "plus",
-  "reload",
-  "remove",
-  "sort",
-  "star",
-  "trash",
-  "upload",
-  "arrow-fill-right",
-  "arrow-fill-left",
-  "chevron-right",
-  "chevron-left",
-  "menu-expand",
-  "menu-collapse",
-  "menu-default",
-  "external",
-  "wallet",
-  "workspace",
-];
-
-export const TAG_THUMBS = [
-  "",
-  "https://avatars3.githubusercontent.com/u/1887591?s=400&u=ba45b6433752099210bf180b4448fb82e015c3a8&v=4",
-];
-
-export const TAG_SKINS = ["pale", "vivid"];
-
-export const TAG_COLORS = [
-  "mineral",
-  "teal",
-  "purple",
-  "green",
-  "red",
-  "yellow",
-  "cyan",
-  "orange",
-  "pink",
-];
